pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pricer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "aiohttp>=3.10",
  "python-dotenv>=1.0",
  "motor>=3.4",          # async MongoDB driver
  "pymongo>=4.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E","F","I","UP","B","BLE","SIM","ASYNC","RUF"]
lint.ignore = ["E501"]
format.quote-style = "double"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
